Store integer attribute values as int64_t to allow out-of-range timestamps
Setting an IfcInteger or IfcTimeStamp typed attribute (for example
IfcOwnerHistory.CreationDate) to a value outside the signed 32-bit range
raised "OverflowError: in method 'entity_instance_setArgumentAsInt',
argument 3 of type 'int'". Any Unix timestamp before 1901-12-13 or after
2038-01-19 fell outside that range, so pre-1901 and post-2038 dates could
not be authored at all.

The scalar integer slot of the attribute variant was a 32-bit int, and the
SWIG setter took a 32-bit int, so a wider Python value overflowed at the
wrapper boundary. Widening only the Python side would silently wrap the
value on file write, which is worse, so the storage, parser and serializer
all need to agree on 64-bit.

This widens just the scalar integer path to int64_t:

* The scalar slot of the attribute variant becomes int64_t. Aggregates of
  integers (Argument_AGGREGATE_OF_INT) stay 32-bit, and instance names and
  reference identifiers stay 32-bit, since neither is the value that
  overflows here.
* IfcBaseClass::set_attribute_value promotes the schema generated int to
  int64_t at a single choke point, so the thousands of generated int
  setters keep compiling unchanged.
* AttributeValue gains an operator int64_t, the STEP text writer and the
  size visitor take int64_t, and the SWIG setArgumentAsInt takes int64_t.
* The STEP lexer parses integer literals with strtoll into a 64-bit token
  field, and TokenFunc::asInt returns int64_t, so values survive a full
  file round trip. The aggregate parser narrows back to int for the
  32-bit integer aggregate slot.
* pythonize gains an int64_t overload (PyLong_FromLongLong) so reading a
  value back never truncates.

Verified in a build (IFC2X3 and IFC4): pre-1901, post-2038, both 32-bit
boundaries and a 9e12 value round trip exactly in memory and through STEP
text serialization, ordinary in-range integers and integer aggregates are
unaffected, and the entity attribute test suite passes.

Fixes #3058

+++ src/ifcparse/IfcParse.h
@@ -80,7 +80,7 @@ class IFC_PARSE_API TokenFunc {
     /// Returns whether the token can be interpreted as a binary type
     static bool isBinary(const Token& token);
     /// Returns the token interpreted as an integer
-    static int asInt(const Token& token);
+    static int64_t asInt(const Token& token);
     /// Returns the token interpreted as an identifier
     static int asIdentifier(const Token& token);
     /// Returns the token interpreted as an boolean (.T. or .F.)
